:memo: Docs: Add TypeScript types for this lib
