+++ package.json
@@ -2,8 +2,8 @@
   "name": "mediascan",
   "version": "1.1.3",
   "description": "A scanner for media files that follows a user-provided naming convention",
-  "main": "./lib/MediaScan.js",
-  "types": "./lib/MediaScan.d.ts",
+  "main": "index.js",
+  "types": "index.d.ts",
   "files": [
     "lib/"
   ],
@@ -12,7 +12,6 @@
     "semantic-release": "semantic-release -e ./config/release.config.js",
     "prepare": "npm run compile",
     "clean": "jest --clearCache",
-    "coverage": "cat ./coverage/lcov.info | coveralls",
     "compile": "babel src --out-dir lib --extensions \".ts\" --source-maps both",
     "type-check": "tsc"
   },
