+++ src/utils/utils_functions.ts
@@ -1,7 +1,7 @@
 // Check properties
 import {access, constants as FsConstants} from "fs";
 import PromiseLib from 'bluebird';
-import MediaScanLib from "../declaration";
+import * as MediaScanTypes from "../MediaScanTypes";
 
 export function checkProperties(obj, properties): boolean {
     return properties.every(x => x in obj && obj[x]);
@@ -20,16 +20,16 @@ export function promisifiedAccess(path) : Promise<any> {
 }
 
 // Default implementation to know which category is this file
-export function defaultWhichCategoryFunction(object : MediaScanLib.TPN) : MediaScanLib.Category{
+export function defaultWhichCategoryFunction(object : MediaScanTypes.TPN) : MediaScanTypes.Category{
     // workaround : const string enum aren't compiled correctly with Babel
     return (checkProperties(object, ['season', 'episode']))
-        ? 'TV_SERIES' as MediaScanLib.Category.TV_SERIES_TYPE : 'MOVIES' as MediaScanLib.Category.MOVIES_TYPE;
+        ? 'TV_SERIES' as MediaScanTypes.Category.TV_SERIES_TYPE : 'MOVIES' as MediaScanTypes.Category.MOVIES_TYPE;
 }
 
 // Generic filter for default properties
 export function filterDefaultProperties<T>(propertiesNames : string[],
-                                           search : MediaScanLib.SearchParameters, meetSpecFunction : (value) => boolean,
-                                           transformFunction : (key : string, value) => MediaScanLib.filterTuple<T> ) : MediaScanLib.filterTuple<T>[]  {
+                                           search : MediaScanTypes.SearchParameters, meetSpecFunction : (value) => boolean,
+                                           transformFunction : (key : string, value) => MediaScanTypes.filterTuple<T> ) : MediaScanTypes.filterTuple<T>[]  {
     return propertiesNames.reduce( (acc, currentProperty) => {
         if (meetSpecFunction(search[currentProperty]))
             acc.push(transformFunction(currentProperty, search[currentProperty]));
